# docs/api/openapi-sketch.yaml
openapi: 3.0.3
info:
  title: AEO MVP API
  description: |
    Backend-only Answer Engine Optimization service.
    Jobs analyze one public site (or demo fixtures) and emit a machine-readable report.
    Visibility experiments are controlled samples (protocol vis-exp-v1), not engine rankings.
  version: 0.1.0

servers:
  - url: http://127.0.0.1:8000
    description: Local MVP

tags:
  - name: health
  - name: jobs

paths:
  /health:
    get:
      tags: [health]
      summary: Liveness probe
      operationId: getHealth
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok

  /api/v1/jobs:
    post:
      tags: [jobs]
      summary: Create an analysis job
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
            examples:
              live:
                value:
                  url: https://example.com/
                  demo_mode: false
                  options:
                    max_pages: 25
                    max_depth: 2
                    runs_per_prompt: 3
                    provider: auto
                    category: answer engine optimization
              demo:
                value:
                  url: https://demo.example/
                  demo_mode: true
                  options:
                    provider: demo
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobCreatedResponse"
        "400":
          description: Invalid URL or options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/jobs/{job_id}:
    get:
      tags: [jobs]
      summary: Get job status and score summary
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
        "404":
          description: Unknown job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/jobs/{job_id}/report:
    get:
      tags: [jobs]
      summary: Get full machine-readable report
      operationId: getJobReport
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Report (completed jobs)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "409":
          description: Job not completed yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unknown job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/jobs/{job_id}/pages:
    get:
      tags: [jobs]
      summary: List crawled pages for a job
      operationId: getJobPages
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Page list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagesResponse"
        "404":
          description: Unknown job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/content-optimization:
    post:
      tags: [jobs]
      summary: Grounded content optimization (Phase 5)
      description: |
        Page intelligence → content gaps → brief → optimized draft.
        Accepts job_id/page_id, SSRF-safe source_url, or offline html.
        Rejects free-form topic generation. paid_llm_opt_in defaults false.
      operationId: contentOptimization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                job_id: { type: string }
                page_id: { type: string }
                source_url: { type: string, format: uri }
                html: { type: string }
                url: { type: string, format: uri }
                queryset: { type: object }
                site_profile: { type: object }
                config: { type: object }
                paid_llm_opt_in: { type: boolean, default: false }
            examples:
              offline:
                value:
                  html: "<html><body><h1>Demo</h1></body></html>"
                  url: https://demo.example/
                  paid_llm_opt_in: false
      responses:
        "200":
          description: Optimization result
          content:
            application/json:
              schema:
                type: object
                required: [page_intelligence, gap_report, brief, draft]
                properties:
                  page_intelligence: { type: object }
                  gap_report: { type: object }
                  brief: { type: object }
                  draft: { type: object }
                  paid_retrieval: { type: boolean }
                  paid_llm: { type: boolean }
        "400":
          description: Validation or SSRF rejection
        "422":
          description: Malformed body (e.g. topic-only generate)

components:
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]

    CreateJobRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: Public http(s) base URL (one site per job)
        demo_mode:
          type: boolean
          default: false
        options:
          $ref: "#/components/schemas/JobOptions"

    JobOptions:
      type: object
      properties:
        max_pages:
          type: integer
          minimum: 1
          maximum: 25
          default: 25
        max_depth:
          type: integer
          minimum: 0
          maximum: 2
          default: 2
        runs_per_prompt:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
        provider:
          type: string
          enum: [auto, demo, openai_compatible]
          default: auto
        category:
          type: string
          description: Category phrase for prompt-set-v1 templates

    JobCreatedResponse:
      type: object
      required: [id, base_url, status, demo_mode, created_at, links]
      properties:
        id:
          type: string
          format: uuid
        base_url:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        demo_mode:
          type: boolean
        created_at:
          type: string
          format: date-time
        links:
          type: object
          properties:
            self: { type: string }
            report: { type: string }
            pages: { type: string }

    JobStatus:
      type: string
      enum:
        - pending
        - crawling
        - analyzing
        - scoring
        - experimenting
        - synthesizing
        - completed
        - failed

    JobStatusResponse:
      type: object
      required: [id, base_url, status, demo_mode, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        base_url: { type: string }
        status: { $ref: "#/components/schemas/JobStatus" }
        demo_mode: { type: boolean }
        error_message:
          type: string
          nullable: true
        health_score:
          type: number
          nullable: true
          minimum: 0
          maximum: 100
        health_formula_version:
          type: string
          nullable: true
          example: health-v1
        experiment_protocol_version:
          type: string
          nullable: true
          example: vis-exp-v1
        component_scores:
          type: object
          nullable: true
          additionalProperties:
            type: number
          properties:
            technical: { type: number }
            content: { type: number }
            entity: { type: number }
            structured_data: { type: number }
            answerability: { type: number }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        completed_at:
          type: string
          format: date-time
          nullable: true

    ScoreValue:
      type: object
      required: [value, provenance]
      properties:
        value: { type: number }
        provenance:
          type: string
          enum: [derived_metric, api_observation, llm_assessment, estimate, synthetic_demo]
        formula_version: { type: string }
        numerator: { type: integer }
        denominator: { type: integer }

    Report:
      type: object
      required:
        - job_id
        - base_url
        - demo_mode
        - status
        - methodology
        - caveats
        - scores
        - experiment
        - findings
        - recommendations
        - pages_crawled
        - emitted_at
      properties:
        job_id: { type: string, format: uuid }
        base_url: { type: string }
        demo_mode: { type: boolean }
        status: { $ref: "#/components/schemas/JobStatus" }
        methodology:
          type: object
          properties:
            health_formula_version: { type: string, example: health-v1 }
            experiment_protocol_version: { type: string, example: vis-exp-v1 }
            prompt_set_id: { type: string, example: prompt-set-v1 }
            crawl:
              type: object
              properties:
                max_pages: { type: integer }
                max_depth: { type: integer }
                user_agent: { type: string }
        caveats:
          type: array
          items: { type: string }
        scores:
          type: object
          properties:
            aeo_health: { $ref: "#/components/schemas/ScoreValue" }
            technical: { $ref: "#/components/schemas/ScoreValue" }
            content: { $ref: "#/components/schemas/ScoreValue" }
            entity: { $ref: "#/components/schemas/ScoreValue" }
            structured_data: { $ref: "#/components/schemas/ScoreValue" }
            answerability: { $ref: "#/components/schemas/ScoreValue" }
        experiment:
          type: object
          properties:
            provider_name: { type: string }
            model_id: { type: string, nullable: true }
            observations_count: { type: integer }
            ai_mention_rate: { $ref: "#/components/schemas/ScoreValue" }
            ai_citation_rate: { $ref: "#/components/schemas/ScoreValue" }
            query_coverage: { $ref: "#/components/schemas/ScoreValue" }
        findings:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              category: { type: string }
              title: { type: string }
              summary: { type: string }
              severity: { type: string }
              evidence_ids:
                type: array
                items: { type: string }
        recommendations:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              code: { type: string }
              title: { type: string }
              rationale: { type: string }
              effort: { type: string, enum: [S, M, L] }
              impact: { type: number }
              priority_score: { type: number }
              rank: { type: integer }
              evidence_ids:
                type: array
                items: { type: string }
              finding_ids:
                type: array
                items: { type: string }
        pages_crawled: { type: integer }
        emitted_at: { type: string, format: date-time }

    PagesResponse:
      type: object
      required: [job_id, count, pages]
      properties:
        job_id: { type: string, format: uuid }
        count: { type: integer }
        pages:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              url: { type: string }
              final_url: { type: string, nullable: true }
              depth: { type: integer }
              status_code: { type: integer, nullable: true }
              title: { type: string, nullable: true }
              fetch_error: { type: string, nullable: true }

    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string